pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Agency MBS trading desk analytics platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0",
    "openai>=1.30",
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "diskcache>=5.6",
    "plotly>=5.0",
    "python-dateutil>=2.8",
    "scikit-learn>=1.3",
    "duckdb>=1.0",
    "weave>=0.50",
    "apscheduler>=3.10",
    "openai-agents>=0.0.9",
    "markdown>=3.10.2",
    "aiofiles>=23.0",
    "pyyaml>=6.0",
    "ldap3>=2.9",
    "uvicorn>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.uv]
package = false
